# swagger-specs.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "ChatGPT API",
    "version": "1.0.0",
    "description": "ChatGPT API文档，包含流式和非流式接口",
    "contact": {
      "name": "开发团队"
    }
  },
  "servers": [
    {
      "url": "http://localhost:3001",
      "description": "开发服务器"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "JWT认证，在请求头中添加：Authorization: Bearer {token}"
      }
    }
  },
  "security": [
    {
      "bearerAuth": []
    }
  ],
  "paths": {
    "/api/knowledge/upload": {
      "post": {
        "tags": [
          "知识库"
        ],
        "summary": "上传并向量化文档",
        "description": "上传PDF、TXT、DOC或DOCX文件，解析内容并添加到用户的知识库中",
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "properties": {
                  "file": {
                    "type": "string",
                    "format": "binary",
                    "description": "要上传的文档文件"
                  },
                  "userId": {
                    "type": "string",
                    "description": "用户ID"
                  },
                  "documentName": {
                    "type": "string",
                    "description": "文档名称"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "文档上传并向量化成功",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean"
                    },
                    "message": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "请求参数错误"
          },
          "500": {
            "description": "服务器错误"
          }
        }
      }
    },
    "/api/knowledge/query": {
      "post": {
        "tags": [
          "知识库"
        ],
        "summary": "查询知识库",
        "description": "使用自然语言查询用户的知识库",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "query",
                  "userId"
                ],
                "properties": {
                  "query": {
                    "type": "string",
                    "description": "查询文本"
                  },
                  "userId": {
                    "type": "string",
                    "description": "用户ID"
                  },
                  "topK": {
                    "type": "integer",
                    "description": "返回的最相关结果数量",
                    "default": 5
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "查询成功",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "results": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "content": {
                            "type": "string"
                          },
                          "metadata": {
                            "type": "object"
                          },
                          "score": {
                            "type": "number"
                          }
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "请求参数错误"
          },
          "500": {
            "description": "服务器错误"
          }
        }
      }
    },
    "/api/knowledge/chat": {
      "post": {
        "tags": [
          "知识库"
        ],
        "summary": "与知识库聊天",
        "description": "使用用户的知识库进行聊天，返回非流式响应",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "message",
                  "userId"
                ],
                "properties": {
                  "message": {
                    "type": "string",
                    "description": "用户消息"
                  },
                  "userId": {
                    "type": "string",
                    "description": "用户ID"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "聊天成功",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "answer": {
                      "type": "string"
                    },
                    "sources": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "content": {
                            "type": "string"
                          },
                          "metadata": {
                            "type": "object"
                          }
                        }
                      }
                    },
                    "reply": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "请求参数错误"
          },
          "500": {
            "description": "服务器错误"
          }
        }
      }
    },
    "/api/knowledge/stream-chat": {
      "post": {
        "tags": [
          "知识库"
        ],
        "summary": "与知识库流式聊天",
        "description": "使用用户的知识库进行聊天，返回流式响应",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "message",
                  "userId"
                ],
                "properties": {
                  "message": {
                    "type": "string",
                    "description": "用户消息"
                  },
                  "userId": {
                    "type": "string",
                    "description": "用户ID"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "返回 SSE 流",
            "content": {
              "text/event-stream": {
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "400": {
            "description": "请求参数错误"
          },
          "500": {
            "description": "服务器错误"
          }
        }
      }
    },
    "/api/knowledge/delete": {
      "delete": {
        "tags": [
          "知识库"
        ],
        "summary": "删除用户知识库",
        "description": "删除指定用户的知识库内容",
        "parameters": [
          {
            "in": "query",
            "name": "documentName",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "可选，指定要删除的文档名称"
          }
        ],
        "responses": {
          "200": {
            "description": "删除成功",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean"
                    },
                    "message": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "请求参数错误"
          },
          "401": {
            "description": "未认证"
          },
          "500": {
            "description": "服务器错误"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
...
  },
...
}